This can create DB, Tables, and some data
This contains the .sql files to create database, create tables, and insert mock data into the database. Added the .csv that the mock data comes from and insertgen.py that reads the .csv files in the csvFiles directory and creates a create_{TableName}.sql for each of the .csv's with proper inserts.